requirements.txt
@"
pyyaml>=5.1
pytest>=7.0
"@ | Out-File -FilePath "requirements.txt" -Encoding UTF8